refactor: job plugin in example
