+++ example/src/plugins/job-ui-single/index.tsx
@@ -1,26 +1,9 @@
 import { TPlugin } from '@development-framework/dm-core'
+import { JobPlugin } from './pages/JobPlugin'
 
-//***********************************************************
-//Forms
-//import { SingleObjectForm} from './marmo/forms/SingleObjectForm'
-//***********************************************************
-
-//***********************************************************
-//views
-import { Jobs } from './pages/JobRunner'
-
-//import { SignalESSForm } from './marmo/containers/forms/SignalESS'
-
-//***********************************************************
-//***********************************************************
 export const plugins: TPlugin[] = [
-  //********************************************************
-  //********************************************************
-
   {
     pluginName: 'signal-job-single',
-    component: Jobs,
+    component: JobPlugin,
   },
-  //********************************************************
-  //********************************************************
 ]
